Adds open definition to the behaviorrun class.
This is important if using a huge number of these objects simultaneously. At some point you need to close H5 files or the OS will throw an error. Once the file is closed, access to the trials table is preserved (it is in memory), but access to streams is not. By using the open() method, it is possible to regain access to the H5 object and to the streams within.

within the code it might look like this:

run_instance.open()
for t in range(len(run_instance.trials)):
    t = run_instance.return_trial(t)
    # do something here. Don't want automatic closing after each trial is returned,
    # as we're going to iterate through them all and don't want to open and close 300 times.
run_instance.close()